+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next release
+
+* Fixed glyph metrics in a case where normalization setting is off.
+
 ## 0.0.2
 
 * CLI tool arguments can now be specified in yaml config (thanks for suggestion @dricholm).